rename routes.go to http.go and move grpc into server
